Update urdf-viz to 0.37
