+++ arci-urdf-viz/Cargo.toml
@@ -29,4 +29,4 @@ easy-ext = "1"
 flaky_test = "0.1"
 portpicker = "0.1"
 tokio = { version = "1.0", features = ["full"] }
-urdf-viz = "0.36"
+urdf-viz = "0.37"
