+++ arci-urdf-viz/tests/test_web.rs
@@ -1,38 +1,12 @@
+mod util;
+
 use std::{f64, time::Duration};
 
 use arci::{nalgebra, Localization, MoveBase, Navigation};
 use arci_urdf_viz::*;
 use assert_approx_eq::assert_approx_eq;
 use urdf_viz::WebServer;
-use url::Url;
-
-fn port_and_url() -> (u16, Url) {
-    let port = portpicker::pick_unused_port().expect("No ports free");
-    let url = Url::parse(&format!("http://127.0.0.1:{port}")).unwrap();
-    (port, url)
-}
-
-#[easy_ext::ext]
-impl WebServer {
-    fn start_background(self) {
-        let handle = self.handle();
-        std::thread::spawn(move || self.start());
-        std::thread::spawn(move || loop {
-            if let Some(positions) = handle.take_target_joint_positions() {
-                *handle.current_joint_positions() = positions;
-            }
-            if let Some(origin) = handle.pop_target_object_origin() {
-                if origin.id == urdf_viz::ROBOT_OBJECT_ID {
-                    *handle.current_robot_origin() = urdf_viz::RobotOrigin {
-                        position: origin.position,
-                        quaternion: origin.quaternion,
-                    };
-                }
-            }
-        });
-        std::thread::sleep(Duration::from_secs(1)); // Wait for web server to start.
-    }
-}
+use util::*;
 
 #[test]
 fn test_set_get_vel() {
