+++ openrr-planner/Cargo.toml
@@ -33,4 +33,4 @@ assert_approx_eq = "1.1"
 clap = { version = "3.1", features = ["derive"] }
 nalgebra = "0.30"
 tracing-subscriber = "0.3"
-urdf-viz = "0.36"
+urdf-viz = "0.37"
